+++ src/config/config-set.ts
@@ -712,7 +712,7 @@ export class ConfigSet {
     resolvedConfigFile?: string | null,
     noProject?: boolean | null,
   ): ParsedCommandLine {
-    let config = { compilerOptions: {}, include: [] }
+    let config = { compilerOptions: {} }
     let basePath = normalizeSlashes(this.rootDir)
     let configFileName: string | undefined
     const ts = this.compilerModule
@@ -741,11 +741,6 @@ export class ConfigSet {
       ...config.compilerOptions,
       ...compilerOptions,
     }
-    /**
-     * Always set include to empty array so fileNames after parseJsonConfigFileContent only contains the least minimum initial
-     * files to utilize LanguageService incremental feature
-     */
-    config.include = []
 
     // parse json, merge config extending others, ...
     const result = ts.parseJsonConfigFileContent(config, ts.sys, basePath, undefined, configFileName)